requirements.txt
pyinstaller>=6.12.0
pystray>=0.19.5
pillow>=11.3.0
